feat(DataTable): Add filtering logic based on active tab state
- Integrated useTabState hook to retrieve the current active tab.
- Implemented filtering logic within the DataTable component to display tasks according to the selected tab (All, Verified, Completed, Overdue).
- Retained tasks as props, ensuring a modular design for better code maintainability.
