+++ my-app/components/task-components/Tasktabs.tsx
@@ -1,3 +1,4 @@
+'use client'
 import * as React from 'react'
 import Tabs from '@mui/material/Tabs'
 import Tab from '@mui/material/Tab'
